ordering fix
